Init rail commit

+++ include/freerdp/window.h
@@ -135,7 +135,6 @@
 #define WS_EX_TRANSPARENT		0x00000020
 #define WS_EX_WINDOWEDGE		0x00000100
 #endif
-
 /**
  * This is a custom extended window style used by XRDP
  * instructing the client to use local window decorations
@@ -188,13 +187,13 @@ struct _WINDOW_STATE_ORDER
 	UINT32 ownerWindowId;
 	UINT32 style;
 	UINT32 extendedStyle;
-	UINT32 showState;
+	UINT8 showState;
 	RAIL_UNICODE_STRING titleInfo;
 	INT32 clientOffsetX;
 	INT32 clientOffsetY;
 	UINT32 clientAreaWidth;
 	UINT32 clientAreaHeight;
-	UINT32 RPContent;
+	UINT8 RPContent;
 	UINT32 rootParentHandle;
 	INT32 windowOffsetX;
 	INT32 windowOffsetY;
